+++ src/main/java/at/hannibal2/skyhanni/data/HypixelData.kt
@@ -2,12 +2,10 @@ package at.hannibal2.skyhanni.data
 
 import at.hannibal2.skyhanni.events.LocationChangeEvent
 import at.hannibal2.skyhanni.events.LorenzChatEvent
-import at.hannibal2.skyhanni.events.PacketEvent
 import at.hannibal2.skyhanni.events.ProfileJoinEvent
 import at.hannibal2.skyhanni.utils.StringUtils.removeColor
 import at.hannibal2.skyhanni.utils.TabListUtils
 import net.minecraft.client.Minecraft
-import net.minecraft.network.play.server.S38PacketPlayerListItem
 import net.minecraftforge.event.world.WorldEvent
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import net.minecraftforge.fml.common.gameevent.TickEvent
@@ -18,7 +16,6 @@ class HypixelData {
     companion object {
         var hypixel = false
         var skyblock = false
-        var dungeon = false
         var mode: String = ""
     }
 
@@ -33,31 +30,14 @@ class HypixelData {
     val areaRegex = Regex("§r§b§l(?<area>[\\w]+): §r§7(?<loc>[\\w ]+)§r")
 
     @SubscribeEvent
-    fun onTabUpdate(event: PacketEvent.ReceiveEvent) {
-        if (dungeon || !hypixel || event.packet !is S38PacketPlayerListItem ||
-            (event.packet.action != S38PacketPlayerListItem.Action.UPDATE_DISPLAY_NAME &&
-                    event.packet.action != S38PacketPlayerListItem.Action.ADD_PLAYER)
-        ) return
-        event.packet.entries.forEach { playerData ->
-            val name = playerData?.displayName?.formattedText ?: playerData?.profile?.name ?: return@forEach
-            areaRegex.matchEntire(name)?.let { result ->
-                dungeon = skyblock && result.groups["area"]?.value == "Dungeon"
-                return@forEach
-            }
-        }
-    }
-
-    @SubscribeEvent
     fun onWorldChange(event: WorldEvent.Load) {
         skyblock = false
-        dungeon = false
     }
 
     @SubscribeEvent
     fun onDisconnect(event: FMLNetworkEvent.ClientDisconnectionFromServerEvent) {
         hypixel = false
         skyblock = false
-        dungeon = false
     }
 
     @SubscribeEvent